tri-naryTree
Implement insert and delete in a tri-nary tree.  A tri-nary tree is much
like a binary
tree but with three child nodes for each parent instead of two -- with
the left node being values
less than the parent, the right node values greater than the parent, and
the middle nodes values
equal to the parent.